Remove Server#requestStringValuesFromDatabase
This method was very similar to requestStringValueFromDatabase and the
only advantage it had over it was that you could retrieve multiple
values in one request instead of making multiple requests. This
advantage is currently not worth keeping the method for.

Associated methods, such as the private method cacheInitialValues() have
been removed as well.